# api.yml
openapi: 3.0.1
info:
  title: attack_tree_modeler
  description: 'API specification for the attack_tree_modeler API service.'
  contact:
    email: na@example.com
  version: 1.0.0
...
servers:
- url: https://riskytrees.com/api/v1
paths:
  /auth/login:
    post:
      tags:
      - auth
      summary: Logins a new user and returns a session token
      operationId: loginUser
      requestBody:
        description: User details that you want to login with. One day this will be a SAML callback.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUser'
        required: true
      responses:
        200:
          description: User logged in succesfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
...
  /projects:
    get:
      tags:
      - project
      summary: Gets a list of projects you have access to
      operationId: getProjects
      responses:
        200:
          description: Got list of projects successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectsListResponse'
    post:
      tags:
      - project
      summary: Creates a new empty project
      operationId: createProject
      requestBody:
        description: Initial details about the project
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProject'
        required: true
      responses:
        200:
          description: Project created succesfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateProjectResponse'
  /projects/{id}/trees:
    get:
      tags:
      - project
      summary: Gets a list of trees relevant to this project
      operationId: getTrees
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        200:
          description: Got list of trees successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTreesResponse'
    post:
      tags:
      - project
      summary: Creates a new empty tree
      operationId: createTree
      requestBody:
        description: Initial details about the tree
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTree'
        required: true
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        200:
          description: Tree created succesfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTreeResponse'
  /nodes/{id}:
    get:
      tags:
      - nodes
      summary: Gets a specific node and some relevant metadata
      operationId: getNode
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        200:
          description: Got node details succesfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetNodeResponse'
  /projects/{projectId}/trees/{treeId}:
    get:
      tags:
      - project
      summary: View details of a single tree
      operationId: getTree
      parameters:
        - in: path
          name: projectId
          schema:
            type: integer
          required: true
        - in: path
          name: treeId
          schema:
            type: integer
          required: true
      responses:
        200:
          description: Got tree successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeResponse'
    put:
      tags:
      - project
      summary: Updates a tree to match the specified payload.
      operationId: updateTree
      parameters:
        - in: path
          name: projectId
          schema:
            type: integer
          required: true
        - in: path
          name: treeId
          schema:
            type: integer
          required: true
      requestBody:
        description: Desired tree contents
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FullTreeData'
      responses:
        200:
          description: Got tree successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeResponse'
  /projects/{projectId}/trees/{treeId}/dag/down:
    get:
      tags:
      - project
      summary: Returns a DAG-like representation of the relationship between this tree and the subtrees of it
      operationId: getTreeDagDown
      parameters:
        - in: path
          name: projectId
          schema:
            type: integer
          required: true
        - in: path
          name: treeId
          schema:
            type: integer
          required: true
      responses:
        200:
          description: Got DAG successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeDAGResponse'
  /projects/{projectId}/model:
    get:
      tags:
      - project
      summary: View selected model for a project
      operationId: getProjectModel
      parameters:
        - in: path
          name: projectId
          schema:
            type: integer
          required: true
      responses:
        200:
          description: Got model successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelectedModelResponse'
    put:
      tags:
        - project
      summary: Change selected model for a project
      operationId: changeProjectModel
      parameters:
        - in: path
          name: projectId
          schema:
            type: integer
          required: true
      requestBody:
        description: Desired model
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectedModel'
      responses:
        200:
          description: Changed model successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelectedModelResponse'
...
  /projects/{projectId}/configs:
    get:
      tags:
      - project
      summary: View config options of a project
      operationId: getProjectConfigList
      parameters:
        - in: path
          name: projectId
          schema:
            type: integer
          required: true
      responses:
        200:
          description: Got config list successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectConfigListResponse'
    post:
      tags:
      - project
      summary: Add a new config of a project
      operationId: addProjectConfig
      parameters:
        - in: path
          name: projectId
          schema:
            type: integer
          required: true
      requestBody:
        description: Desired config
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectConfigPayload'
        required: true
      responses:
        200:
          description: Changed config successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectConfigResponse'
...
components:
  schemas:
    RegisterUser:
      type: object
      properties:
        email:
          type: string
          format: email
    CreateProject:
      type: object
      properties:
        title:
          type: string
    AuthLoginResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
      type: object
      properties:
        result:
          type: object
          properties:
            sessionToken:
              type: string
    TreeDagItem:
      type: object
      properties:
        id:
          type: string
        children:
          type: array
          items:
            $ref: '#/components/schemas/TreeDagItem'
    TreeDAGResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
      type: object
      properties:
        result:
          type: object
          properties:
            root: 
              $ref: '#/components/schemas/TreeDagItem'
    ProjectConfigListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
      type: object
      properties:
        result:
          type: object
          properties:
            ids:
              type: array
              items:
                type: string
...
    ProjectConfigResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
      type: object
      properties:
        result:
          properties:
            selectedConfig:
              type: string
            attributes:
              type: object
    ProjectConfigPayload:
      type: object
      properties:
        attributes:
          type: object
...
    GetNodeResponse:
      allOf:
      - $ref: '#/components/schemas/ApiResponse'
      type: object
      properties:
        result:
          type: object
          properties:
            treeId:
              type: string
              example: "uuid string"
    SelectedModel:
      type: object
      properties:
        modelId:
          type: string
    SelectedModelResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
      type: object
      properties:
        result:
          $ref: '#/components/schemas/SelectedModel'
    CreateProjectResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
      type: object
      properties:
        result:
          type: object
          properties:
            id:
              type: string
            title:
              type: string
    ProjectsListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
      type: object
      properties:
        result:
          type: object
          properties:
            projects:
              type: array
              items:
                type: object
                properties:
                  projectId:
                    type: string
                    example: 61aewfhewfh1u124
                  name:
                    type: string
                    example: Project Name
...
    FullNodeData:
      allOf:
        - $ref: '#/components/schemas/CreateNode'
      type: object
      properties:
        id:
          type: string
        children:
          type: array
          items:
            type: integer
    CreateNode:
      type: object
      required:
        - title
        - parents
      properties:
        title:
          type: string
        description:
          type: string
        modelAttributes:
          type: object
        conditionAttribute:
          type: string
...
    ListTreesResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
      type: object
      properties:
        result:
          type: object
          properties:
            trees:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  title:
                    type: string
    FullTreeData:
      allOf:
        - $ref: '#/components/schemas/CreateTree'
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/FullNodeData'
        rootNodeId:
          type: string
    TreeResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
      type: object
      properties:
        result:
          $ref: '#/components/schemas/FullTreeData'
    CreateTree:
      type: object
      properties:
        title:
          type: string
    CreateTreeResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
      type: object
      properties:
        result:
          type: object
          properties:
            id:
              type: string
            title:
              type: string
    ApiResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
        result:
          type: object