# riskytrees/src/main.rs
#![feature(proc_macro_hygiene, decl_macro)]

#[macro_use]
extern crate rocket;
extern crate rocket_cors;

use std::collections::HashSet;
use database::get_tree_from_node_id;
use models::ApiProjectConfigListResponseResult;
use mongodb::bson::doc;
use rocket_contrib::json::Json;

use rocket::{http::Method, Config, config::Environment};

use rocket_cors::{
    AllowedHeaders, AllowedOrigins, Cors, CorsOptions
};

mod constants;
mod database;
mod errors;
mod helpers;
mod models;

#[cfg(test)]
mod tests;

fn make_cors() -> Cors {
    let mut origins = HashSet::new();

    origins.insert("*".to_string());

    /*let allowed_origins = AllowedOrigins::Some(rocket_cors::Origins{
        allow_null: true,
        exact: Some(origins),
        regex: None
    });*/

    let allowed_origins = AllowedOrigins::All;

    CorsOptions {
        allowed_origins,
        allowed_methods: vec![Method::Get, Method::Put].into_iter().map(From::from).collect(),
        allowed_headers: AllowedHeaders::some(&[
            "Authorization",
            "Accept",
            "Access-Control-Allow-Origin",
        ]),
        allow_credentials: true,
        ..Default::default()
    }
    .to_cors()
    .expect("error while building CORS")
}


#[get("/")]
fn index() -> &'static str {
    let db_client = database::get_instance();
    match db_client {
        Ok(client) => {
            database::new_user(client, "Test".to_string());
            "Hello, world!"
        }
        Err(e) => "Failed to create user",
    }
}

#[post("/auth/login", data = "<body>")]
fn auth_login(body: Json<models::ApiRegisterUser>) -> Json<models::ApiAuthLoginResponse> {
    let db_client = database::get_instance();
    match db_client {
        Ok(client) => {
            if database::get_user(client.to_owned(), body.email.to_owned()).is_none() {
                database::new_user(client, body.email.to_owned());

                Json(models::ApiAuthLoginResponse {
                    ok: true,
                    message: "User created and logged in succesfully".to_owned(),
                    result: Some(models::AuthLoginResponseResult {
                        sessionToken: "testtoken".to_owned(),
                    }),
                })
            } else {
                Json(models::ApiAuthLoginResponse {
                    ok: true,
                    message: "User logged in succesfully".to_owned(),
                    result: None,
                })
            }
        }
        Err(e) => Json(models::ApiAuthLoginResponse {
            ok: false,
            message: "Could not connect to DB".to_owned(),
            result: None,
        }),
    }
}

#[post("/auth/logout")]
fn auth_logout() -> &'static str {
    "todo"
}

#[get("/projects")]
fn projects_get() -> Json<models::ApiProjectsListResponse> {
    let db_client = database::get_instance();
    match db_client {
        Ok(client) => {
            let ids = database::get_available_project_ids(&client);
            match ids {
                Ok(ids) => Json(models::ApiProjectsListResponse {
                    ok: true,
                    message: "Got projects".to_owned(),
                    result: Some(models::ApiProjectsListResponseResult {
                        projects: database::get_projects_from_ids(ids, &client)
                    }),
                }),
                Err(err) => Json(models::ApiProjectsListResponse {
                    ok: false,
                    message: "Failed to find project ids".to_owned(),
                    result: None,
                })
            }
        }
        Err(e) => Json(models::ApiProjectsListResponse {
            ok: false,
            message: "Could not connect to DB".to_owned(),
            result: None,
        }),
    }
}

#[post("/projects", data = "<body>")]
fn projects_post(body: Json<models::ApiCreateProject>) -> Json<models::ApiCreateProjectResponse> {
    let db_client = database::get_instance();
    match db_client {
        Ok(client) => {
            let project = database::get_project_by_title(client.to_owned(), body.title.to_owned());
            match project {
                Some(project) => Json(models::ApiCreateProjectResponse {
                    ok: true,
                    message: "Tree already exists".to_owned(),
                    result: None,
                }),
                None => match database::new_project(client, body.title.to_owned()) {
                    Ok(new_project_id) => Json(models::ApiCreateProjectResponse {
                        ok: true,
                        message: "Project created succesfully".to_owned(),
                        result: Some(models::CreateProjectResponseResult {
                            title: body.title.to_owned(),
                            id: new_project_id,
                        }),
                    }),
                    Err(err) => Json(models::ApiCreateProjectResponse {
                        ok: false,
                        message: format!("Failed to create project: {}", err),
                        result: None,
                    }),
                },
            }
        }
        Err(e) => Json(models::ApiCreateProjectResponse {
            ok: false,
            message: "Could not connect to DB".to_owned(),
            result: None,
        }),
    }
}

#[post("/projects/<id>/trees", data = "<body>")]
fn projects_trees_post(
    id: String,
    body: Json<models::ApiCreateTree>,
) -> Json<models::ApiCreateTreeResponse> {
    let db_client = database::get_instance();
    match db_client {
        Ok(client) => {
            let project = database::get_project_by_id(&client, id.to_owned());
            match project {
                Some(project) => {
                    // Create tree
                    match database::create_project_tree(
                        client.to_owned(),
                        body.title.to_owned(),
                        id,
                    ) {
                        Ok(db_res) => Json(models::ApiCreateTreeResponse {
                            ok: true,
                            message: "Added tree".to_owned(),
                            result: Some(models::CreateTreeResponseResult {
                                title: body.title.to_owned(),
                                id: db_res,
                            }),
                        }),
                        Err(err) => Json(models::ApiCreateTreeResponse {
                            ok: false,
                            message: format!(
                                "Failed to create new tree and link to project: {}",
                                err
                            ),
                            result: None,
                        }),
                    }
                }
                None => Json(models::ApiCreateTreeResponse {
                    ok: true,
                    message: "Could not find project".to_owned(),
                    result: None,
                }),
            }
        }
        Err(e) => Json(models::ApiCreateTreeResponse {
            ok: false,
            message: "Could not connect to DB".to_owned(),
            result: None,
        }),
    }
}

#[get("/projects/<id>/trees")]
fn projects_trees_get(id: String) -> Json<models::ApiListTreeResponse> {
    let db_client = database::get_instance();
    match db_client {
        Ok(client) => {
            let project: Option<models::Project> =
                database::get_project_by_id(&client, id.to_owned());
            match project {
                Some(project) => {
                    // Get trees
                    let trees = database::get_trees_by_project_id(&client, project.id);

                    match trees {
                        Ok(trees) => Json(models::ApiListTreeResponse {
                            ok: true,
                            message: "Got trees succesfully.".to_owned(),
                            result: Some(models::ListTreeResponseResult { trees }),
                        }),
                        Err(_) => Json(models::ApiListTreeResponse {
                            ok: false,
                            message: "Failed to get projects from db".to_owned(),
                            result: None,
                        }),
                    }
                }
                None => Json(models::ApiListTreeResponse {
                    ok: false,
                    message: "Could not find project".to_owned(),
                    result: None,
                }),
            }
        }
        Err(e) => Json(models::ApiListTreeResponse {
            ok: false,
            message: "Could not connect to DB".to_owned(),
            result: None,
        }),
    }
}

#[get("/projects/<id>/trees/<tree_id>")]
fn projects_trees_tree_get(id: String, tree_id: String) -> Json<models::ApiTreeResponse> {
    let db_client = database::get_instance();
    match db_client {
        Ok(client) => {
            let project: Option<models::Project> =
                database::get_project_by_id(&client, id.to_owned());
            match project {
                Some(project) => {
                    let tree = database::get_tree_by_id(&client, tree_id.to_owned());
                    match tree {
                        Ok(tree) => {
                            Json(models::ApiTreeResponse {
                                ok: true,
                                message: "Found tree".to_owned(),
                                result: Some(tree),
                            })
                        },
                        Err(err) => {
                            Json(models::ApiTreeResponse {
                                ok: false,
                                message: "Could not find tree using id".to_owned(),
                                result: None,
                            })
                        }
                    }

                }
                None => Json(models::ApiTreeResponse {
                    ok: false,
                    message: "Could not find project".to_owned(),
                    result: None,
                }),
            }
        }
        Err(e) => Json(models::ApiTreeResponse {
            ok: false,
            message: "Could not connect to DB".to_owned(),
            result: None,
        }),
    }
}

#[put("/projects/<id>/trees/<tree_id>", data = "<body>")]
fn projects_trees_tree_put(id: String, tree_id: String, body: Json<models::ApiFullTreeData>) -> Json<models::ApiTreeResponse> {
    let db_client = database::get_instance();
    match db_client {
        Ok(client) => {
            let project: Option<models::Project> =
                database::get_project_by_id(&client, id.to_owned());
            match project {
                Some(project) => {
                    // Update tree and return
                    let tree = database::update_tree_by_id(&client, tree_id.to_owned(), models::ApiFullTreeData {
                        title: body.title.to_owned(),
                        rootNodeId: body.rootNodeId.to_owned(),
                        nodes: body.nodes.clone()
                    });
                    match tree {
                        Ok(tree) => {
                            Json(models::ApiTreeResponse {
                                ok: true,
                                message: "Found tree".to_owned(),
                                result: Some(tree),
                            })
                        },
                        Err(err) => {
                            Json(models::ApiTreeResponse {
                                ok: false,
                                message: "Could not find tree using id".to_owned(),
                                result: None,
                            })
                        }
                    }

                }
                None => Json(models::ApiTreeResponse {
                    ok: false,
                    message: "Could not find project".to_owned(),
                    result: None,
                }),
            }
        }
        Err(e) => Json(models::ApiTreeResponse {
            ok: false,
            message: "Could not connect to DB".to_owned(),
            result: None,
        }),
    }
}


#[get("/projects/<id>/model")]
fn projects_model_get(id: String) -> Json<models::ApiSelectedModelResponse> {
    let db_client = database::get_instance();
    match db_client {
        Ok(client) => {
            let project: Option<models::Project> =
                database::get_project_by_id(&client, id.to_owned());
            match project {
                Some(project) => {
                    Json(models::ApiSelectedModelResponse {
                        ok: true,
                        message: "Found tree".to_owned(),
                        result: Some(models::SelectedModelResult {
                            modelId: project.selected_model.unwrap_or_default(),
                        }),
                    })
                
                }
                None => Json(models::ApiSelectedModelResponse {
                    ok: false,
                    message: "Could not find project".to_owned(),
                    result: None,
                }),
            }
        }
        Err(e) => Json(models::ApiSelectedModelResponse {
            ok: false,
            message: "Could not connect to DB".to_owned(),
            result: None,
        }),
    }
}

#[put("/projects/<id>/model", data = "<body>")]
fn projects_model_put(id: String, body: Json<models::SelectedModelResult>) -> Json<models::ApiSelectedModelResponse> {
    let db_client = database::get_instance();
    match db_client {
        Ok(client) => {
            let project: Option<models::Project> =
                database::get_project_by_id(&client, id.to_owned());
            match project {
                Some(project) => {
                    match database::update_project_model(client, id.to_owned(), body.modelId.to_owned()) {
                        Ok(_) => {
                            Json(models::ApiSelectedModelResponse {
                                ok: true,
                                message: "Updated project".to_owned(),
                                result: Some(models::SelectedModelResult {
                                    modelId: body.modelId.to_owned(),
                                }),
                            })
                        },
                        Err(err) => {
                            Json(models::ApiSelectedModelResponse {
                                ok: false,
                                message: "Could not update project".to_owned(),
                                result: None,
                            })
                        }
                    }
                
                }
                None => Json(models::ApiSelectedModelResponse {
                    ok: false,
                    message: "Could not find project".to_owned(),
                    result: None,
                }),
            }
        }
        Err(e) => Json(models::ApiSelectedModelResponse {
            ok: false,
            message: "Could not connect to DB".to_owned(),
            result: None,
        }),
    }
}

#[get("/models")]
fn models_get() -> Json<models::ApiListModelResponse> {
    let db_client = database::get_instance();

    let model_list = vec![
        models::ListModelResponseItem {
            id: "b9ff54e0-37cf-41d4-80ea-f3a9b1e3af74".to_owned(), // V4 UUID
            title: "Attacker Likelihood".to_owned(),
        },
        models::ListModelResponseItem {
            id: "f1644cb9-b2a5-4abb-813f-98d0277e42f2".to_owned(), // V4 UUID
            title: "Risk of Attack".to_owned(),
        },
        models::ListModelResponseItem {
            id: "bf4397f7-93ae-4502-a4a2-397f40f5cc49".to_owned(), // V4 UUID
            title: "EVITA".to_owned(),
        }
    ];

    match db_client {
        Ok(client) => {
            Json(models::ApiListModelResponse {
                ok: false,
                message: "Could not connect to DB".to_owned(),
                result: Some(models::ListModelResult { 
                    models: model_list
                }),
            })
        }
        Err(e) => Json(models::ApiListModelResponse {
            ok: false,
            message: "Could not connect to DB".to_owned(),
            result: None,
        }),
    }
}

#[get("/nodes/<id>")]
fn node_get(id: String) -> Json<models::ApiGetNodeResponse> {
    let db_client = database::get_instance();

    match db_client {
        Ok(client) => {
            match get_tree_from_node_id(id, &client) {
                Ok(res) => Json(res),
                Err(err) => Json(models::ApiGetNodeResponse {
                    ok: false,
                    message: "Could not find node".to_owned(),
                    result: None,
                })
            }
        },
        Err(err) => Json(models::ApiGetNodeResponse {
            ok: false,
            message: "Could not connect to DB".to_owned(),
            result: None,
        })
    }
}

#[get("/projects/<projectId>/trees/<treeId>/dag/down")]
fn projects_trees_tree_dag_down_get(projectId: String, treeId: String) -> Json<models::ApiTreeDagResponse> {
    let db_client = database::get_instance();

    match db_client {
        Ok(client) => {
            let result = models::ApiTreeDagResponseResult {
                root: models::ApiTreeDagItem {
                    id: treeId.clone(),
                    children: database::get_tree_relationships_down(&treeId, &client)
                }
            };

            Json(models::ApiTreeDagResponse {
                ok: true,
                message: "Got relationship".to_string(),
                result: Some(result)
            })
        },
        Err(err) => Json(models::ApiTreeDagResponse {
            ok: false,
            message: "Could not connect to DB".to_owned(),
            result: None,
        })
    }
}

#[get("/projects/<projectId>/configs")]
fn projects_configs_get(projectId: String) -> Json<models::ApiProjectConfigListResponse> {
    let db_client = database::get_instance();

    match db_client {
        Ok(client) => {
            let matching_configs = database::get_configs_for_project(&projectId, &client);

            Json(models::ApiProjectConfigListResponse {
                ok: true,
                message: "Got configs".to_string(),
                result: Some(ApiProjectConfigListResponseResult {
                    ids: matching_configs
                })
            })
        },
        Err(err) => Json(models::ApiProjectConfigListResponse {
            ok: false,
            message: "Could not connect to DB".to_owned(),
            result: None,
        })
    }
}

fn main() {
    let config = Config::build(Environment::Staging)
    .address("0.0.0.0")
    .unwrap();

    rocket::custom(config)
        .mount(
            "/",
            routes![
                index,
                auth_login,
                auth_logout,
                projects_get,
                projects_post,
                projects_trees_post,
                projects_trees_get,
                projects_trees_tree_get,
                projects_trees_tree_put,
                projects_trees_tree_dag_down_get,
                projects_model_get,
                projects_model_put,
                projects_configs_get,
                models_get,
                node_get
            ],
        )
        .attach(make_cors())

        .launch();
}